feat(items): individual items can be edited, from the 'edit items' page, which is then stored to the database, and the user can see these changes in 'all items' list (#28)
* feat(itemDetailsPage): 'Save' button submits the form values

used the form= attribute for this - see stackoverflow answer here: https://stackoverflow.com/a/12567605/16795175 .  had to adjust the css styling of the 'save' button as a result. opted for this option of having the submit button outside of the <form> because I think it's clearer, maintenance-wise, that the save and cancel buttons are in the 'footer'. that's really where it is visually.

* feat(items): create a POST route at /items/:id/update, which the user arrives at when they submit the itemDetailsPage form. log the request.body to see the submitted form.

* refactor(itemDetailsPage): rename the 'name' attribute of input fields to align with field names in database

helps with maintenance, rather than needing to figure out what translates to what field

* feat(items): on click of 'Save' button on the 'Edit item' page, the itemsController's 'updateItemPost' method is called, transforming the data, and then calling the db query 'updateItemById'. with a helper function (generateUpdateQuery), we can use pg library to update the db for that particular item (identified by ID to be safest)

solution to use the generateUpdateQuery helper function is pretty much this stackoverflow answer: https://stackoverflow.com/a/21763631/16795175 . to be safe, i've also avoided potential sql injection by inserting malicious id by using idx + 2, and using a parametised value for the item id in db.updateItemById

* feat(itemsController): once an item has been edited, the user is redirected back to the 'All items' page

* feat(db/queries): items are sorted by id

before this commit, because there wasn't an 'ORDER BY' in the db query, after an item gets edited, it then goes to the bottom of the list, which isn't correct.

* refactor(updateItemPost): use spread operator to isolate the price_dollars property in the request body, for transformation into price_cents for storage in database

* feat(items): weight and price (in cents) can both be floats

this makes sense in practice if we really want to be super precise. having them as integers before isn't really realistic

+++ src/db/populatedb.js
@@ -14,10 +14,10 @@ CREATE TABLE IF NOT EXISTS items (
     id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name VARCHAR(100) NOT NULL,
     sku VARCHAR(24) UNIQUE NOT NULL,
-    size_grams SMALLINT NOT NULL,
+    size_grams DECIMAL NOT NULL,
     roastery VARCHAR(100) NOT NULL,
     description VARCHAR(4000),
-    price_cents SMALLINT NOT NULL,
+    price_cents DECIMAL NOT NULL,
     stock_quantity NUMERIC NOT NULL,
     category_id INTEGER REFERENCES categories(id)
 );
